# infrastructure/openapi/image-service.json
{"openapi":"3.1.0","info":{"title":"Nerdcode Programme Studio - Image Service","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/ready":{"get":{"summary":"Ready","operationId":"ready_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/image-jobs/plan":{"post":{"summary":"Create Execution Plan","operationId":"create_execution_plan_v1_image_jobs_plan_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageJobRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/image-jobs":{"post":{"summary":"Create Job","operationId":"create_job_v1_image_jobs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageJobRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/image-jobs/{reference_number}":{"get":{"summary":"Get Job","operationId":"get_job_v1_image_jobs__reference_number__get","parameters":[{"name":"reference_number","in":"path","required":true,"schema":{"type":"string","title":"Reference Number"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/image-jobs/{reference_number}/execute":{"post":{"summary":"Execute Job","operationId":"execute_job_v1_image_jobs__reference_number__execute_post","parameters":[{"name":"reference_number","in":"path","required":true,"schema":{"type":"string","title":"Reference Number"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/image-jobs/{reference_number}/manifest":{"post":{"summary":"Create Manifest","operationId":"create_manifest_v1_image_jobs__reference_number__manifest_post","parameters":[{"name":"reference_number","in":"path","required":true,"schema":{"type":"string","title":"Reference Number"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/image-jobs/{reference_number}/publish":{"post":{"summary":"Publish Job","operationId":"publish_job_v1_image_jobs__reference_number__publish_post","parameters":[{"name":"reference_number","in":"path","required":true,"schema":{"type":"string","title":"Reference Number"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/image-jobs/{reference_number}/workflow":{"post":{"summary":"Run Workflow","operationId":"run_workflow_v1_image_jobs__reference_number__workflow_post","parameters":[{"name":"reference_number","in":"path","required":true,"schema":{"type":"string","title":"Reference Number"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Workflow","operationId":"get_workflow_v1_image_jobs__reference_number__workflow_get","parameters":[{"name":"reference_number","in":"path","required":true,"schema":{"type":"string","title":"Reference Number"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageJobRequest":{"properties":{"reference_number":{"type":"string","maxLength":13,"minLength":13,"title":"Reference Number"}},"type":"object","required":["reference_number"],"title":"ImageJobRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}